Split up TF utils.go.erb (#3071)
* split up utils file

* add error retry preds to conversion for now
